Extra features, bundled ezpyle
Renamed patch_target to load_and_patch to be more memorable, added a toggle for whether to flash the LED on serial input, added a way to recall the previously typed line of text to in_line.
Also, bundled a slightly modified version of ezpyle to allow for file editing over the serial connection.